chore : token 관련 상수 이동 - constant 폴더 -> token.ts 파일 내부
