--- src/shared/constants/auth.ts
@@ -1,10 +1,0 @@
-export const USER_DATA_KEY = 'user_data';
-
-export const AUTH_ERRORS = {
-  REQUIRED_COOKIE_NOT_FOUND: 'REQUIRED_COOKIE_NOT_FOUND',
-  UNSUPPORTED_JWT: 'UNSUPPORTED_JWT',
-  LOGGED_OUT_USER: 'LOGGED_OUT_USER',
-  INVALID_JWT_SIGNATURE: 'INVALID_JWT_SIGNATURE',
-  EXPIRED_REFRESH_TOKEN: 'EXPIRED_REFRESH_TOKEN',
-  INVALID_REFRESH_TOKEN: 'INVALID_REFRESH_TOKEN',
-} as const;
